n9sort code progress @ FRIB

+++ src/pid.h
@@ -10,7 +10,6 @@
 
 #include <string>
 
-#include "constants.h"
 #include "ZApar.h"
 #include "SortConfig.h"
 
@@ -21,7 +20,7 @@ public:
 	~pid();
 
 	bool getPID(float x, float y);
-	double getMass(int iZ,int iA);
+	double getMass(size_t iZ, size_t iA);
 
 	ZApar** par; // individual banana gates
 	int nlines;  // number of banana gated stored
